scale_tests: rename TieredStorageWithLoadTest->HighThroughputTest
